move linegraph component to another file and restructure file. chart not showing up because data state not updating properly
